stop trying to call JSON->import for JSON::PP
The configuration for JSON::PP was copying the configuration for JSON
v2. This would mostly work, but included a call to JSON->import to
configure the module. Up until now, that import call would just be
ignored if JSON.pm was not loaded, but future versions of perl will
start warning or dying for that.

Instead, make the JSON.pm v2 configuration a copy of JSON::PP, but
wrapping the create_object sub to add the import call.